requirements.txt
strands-agents[a2a,openai]
strands-agents-tools
openai
pydantic
markdown
click
python-dotenv
jsonschema
aiofiles
pyyaml
